--- a/go.mod
+++ b/go.mod
@@ -1,3 +1,3 @@
 module sigs.k8s.io/json
 
-go 1.18
+go 1.21