--- a/internal/golang/encoding/json/bench_test.go
+++ b/internal/golang/encoding/json/bench_test.go
@@ -14,7 +14,6 @@ import (
 	"bytes"
 	"compress/gzip"
 	"fmt"
-	"internal/testenv"
 	"io"
 	"os"
 	"reflect"
@@ -462,9 +461,7 @@ func BenchmarkUnmapped(b *testing.B) {
 func BenchmarkTypeFieldsCache(b *testing.B) {
 	b.ReportAllocs()
 	var maxTypes int = 1e6
-	if testenv.Builder() != "" {
-		maxTypes = 1e3 // restrict cache sizes on builders
-	}
+	maxTypes = 1e3 // restrict cache sizes on builders
 
 	// Dynamically generate many new types.
 	types := make([]reflect.Type, maxTypes)