… internal/golang/encoding/json/encode_test.go
… internal/golang/encoding/json/fold_test.go
… internal/golang/encoding/json/scanner_test.go
… internal/golang/encoding/json/stream_test.go
… internal/golang/encoding/json/tagkey_test.go