First commit is a straight copy of the json/encoding files from stdlib on the go 1.23 release branch

Second commit is a reapply of carry patches to those files

Third commit is a tweak to make the external dependency check work with go 1.23 `go mod graph` behavior