@jpbetz 2024-10-08
I compared the carry patch commit against the "patch:..." commits in the repo history and it all looks right to me. The other changes look good too.

/approve
/lgtm