@liggitt 2024-10-08
oof... sorry... this was still WIP ... it appears the stdlib json byte use may have regressed between 1.19 and 1.23 and I was chasing that down... will follow that up and resolve or revert before bumping in k/k