@liggitt 2024-10-08
```
benchcmp 1.20.14.txt 1.21.13.txt 
benchmark                                old ns/op     new ns/op     delta
BenchmarkUnmarshal/typed_stdlib-10       33821         32411         -4.17%
BenchmarkUnmarshal/typed_stdlib-10       33654         32313         -3.98%
BenchmarkUnmarshal/typed_stdlib-10       33417         32258         -3.47%
BenchmarkUnmarshal/typed_stdlib-10       33361         32598         -2.29%
BenchmarkUnmarshal/typed_stdlib-10       33344         32292         -3.15%
BenchmarkUnmarshal/typed_stdlib-10       33385         32259         -3.37%
BenchmarkUnmarshal/typed_stdlib-10       33198         32223         -2.94%
BenchmarkUnmarshal/typed_stdlib-10       33390         32296         -3.28%
BenchmarkUnmarshal/typed_stdlib-10       34015         32165         -5.44%
BenchmarkUnmarshal/typed_stdlib-10       34059         33392         -1.96%
BenchmarkUnmarshal/untyped_stdlib-10     39088         36741         -6.00%
BenchmarkUnmarshal/untyped_stdlib-10     38622         35609         -7.80%
BenchmarkUnmarshal/untyped_stdlib-10     38615         37185         -3.70%
BenchmarkUnmarshal/untyped_stdlib-10     41221         35654         -13.51%
BenchmarkUnmarshal/untyped_stdlib-10     38330         35452         -7.51%
BenchmarkUnmarshal/untyped_stdlib-10     38662         35238         -8.86%
BenchmarkUnmarshal/untyped_stdlib-10     38069         35582         -6.53%
BenchmarkUnmarshal/untyped_stdlib-10     38133         36464         -4.38%
BenchmarkUnmarshal/untyped_stdlib-10     37744         35498         -5.95%
BenchmarkUnmarshal/untyped_stdlib-10     38408         35433         -7.75%

benchmark                                old allocs     new allocs     delta
BenchmarkUnmarshal/typed_stdlib-10       161            162            +0.62%
BenchmarkUnmarshal/typed_stdlib-10       161            162            +0.62%
BenchmarkUnmarshal/typed_stdlib-10       161            162            +0.62%
BenchmarkUnmarshal/typed_stdlib-10       161            162            +0.62%
BenchmarkUnmarshal/typed_stdlib-10       161            162            +0.62%
BenchmarkUnmarshal/typed_stdlib-10       161            162            +0.62%
BenchmarkUnmarshal/typed_stdlib-10       161            162            +0.62%
BenchmarkUnmarshal/typed_stdlib-10       161            162            +0.62%
BenchmarkUnmarshal/typed_stdlib-10       161            162            +0.62%
BenchmarkUnmarshal/typed_stdlib-10       161            162            +0.62%
BenchmarkUnmarshal/untyped_stdlib-10     517            515            -0.39%
BenchmarkUnmarshal/untyped_stdlib-10     517            515            -0.39%
BenchmarkUnmarshal/untyped_stdlib-10     517            515            -0.39%
BenchmarkUnmarshal/untyped_stdlib-10     516            515            -0.19%
BenchmarkUnmarshal/untyped_stdlib-10     517            515            -0.39%
BenchmarkUnmarshal/untyped_stdlib-10     517            515            -0.39%
BenchmarkUnmarshal/untyped_stdlib-10     517            515            -0.39%
BenchmarkUnmarshal/untyped_stdlib-10     517            515            -0.39%
BenchmarkUnmarshal/untyped_stdlib-10     517            515            -0.39%
BenchmarkUnmarshal/untyped_stdlib-10     517            515            -0.39%

benchmark                                old bytes     new bytes     delta
BenchmarkUnmarshal/typed_stdlib-10       13240         20384         +53.96%
BenchmarkUnmarshal/typed_stdlib-10       13240         20384         +53.96%
BenchmarkUnmarshal/typed_stdlib-10       13240         20384         +53.96%
BenchmarkUnmarshal/typed_stdlib-10       13240         20384         +53.96%
BenchmarkUnmarshal/typed_stdlib-10       13240         20384         +53.96%
BenchmarkUnmarshal/typed_stdlib-10       13240         20384         +53.96%
BenchmarkUnmarshal/typed_stdlib-10       13240         20384         +53.96%
BenchmarkUnmarshal/typed_stdlib-10       13240         20384         +53.96%
BenchmarkUnmarshal/typed_stdlib-10       13240         20384         +53.96%
BenchmarkUnmarshal/typed_stdlib-10       13240         20384         +53.96%
BenchmarkUnmarshal/untyped_stdlib-10     28163         27733         -1.53%
BenchmarkUnmarshal/untyped_stdlib-10     28165         27737         -1.52%
BenchmarkUnmarshal/untyped_stdlib-10     28167         27733         -1.54%
BenchmarkUnmarshal/untyped_stdlib-10     28159         27734         -1.51%
BenchmarkUnmarshal/untyped_stdlib-10     28165         27734         -1.53%
BenchmarkUnmarshal/untyped_stdlib-10     28165         27736         -1.52%
BenchmarkUnmarshal/untyped_stdlib-10     28161         27735         -1.51%
BenchmarkUnmarshal/untyped_stdlib-10     28167         27734         -1.54%
BenchmarkUnmarshal/untyped_stdlib-10     28160         27739         -1.50%
BenchmarkUnmarshal/untyped_stdlib-10     28161         27736         -1.51%
```