@liggitt 2024-10-08
yeah, 50% regression in byte use in typed unmarshals from stdlib 1.21 → 1.22