@liggitt 2024-10-08
full change from 1.20 to 1.21:

git log --format=oneline upstream/release-branch.go1.20..upstream/release-branch.go1.21 -- src/encoding/json/

* https://github.com/golang/go/commit/04c628935d1487632f903e4e0688fb7a34063752 encoding/json: document unsupported float values
* https://github.com/golang/go/commit/2e51f6f25cfc37f1e1d6fd3a90f0570979939a21 encoding/json: make use of Buffer.AvailableBuffer
* https://github.com/golang/go/commit/b9b8cecbfc72168ca03ad586cc2ed52b0e8db409 encoding/json: simplify folded name logic
* https://github.com/golang/go/commit/2de406bb9e26df19a31b5f6111bb221b60964d48 encoding/json: use append for HTMLEscape
* https://github.com/golang/go/commit/cb3e170ae764856a99160ee1a6cf8f04c8f51e5d encoding/json: rely on reflect.Value.SetZero
* https://github.com/golang/go/commit/8367e2dfc72f626cba75be71112d42fdb2ed82f7 encoding/json: remove legacy fuzz.go file
* https://github.com/golang/go/commit/ac27b4dd1d72bd6467bf7bfa8bc77a1522efd4cc encoding/json: rely on reflect.Value.Grow
* https://github.com/golang/go/commit/8e5f56a2e3a027e886d78f36675c275b9c845da0 encoding/json: use append-like operations for encoding
* https://github.com/golang/go/commit/e2f413402527505144beea443078649380e0c545 encoding/json: unify encodeState.string and encodeState.stringBytes
* https://github.com/golang/go/commit/21ff6704bc8efa72abe191263aae938f3c867480 encoding/json: use append for Compact and Indent