@liggitt 2024-10-09
opened a revert of this at https://github.com/kubernetes-sigs/json/pull/20 while I dig into the regression