@liggitt 2024-10-09
bisected to https://github.com/golang/go/commit/ac27b4dd1d72bd6467bf7bfa8bc77a1522efd4cc encoding/json: rely on reflect.Value.Grow