@liggitt 2024-10-09
captured test results at https://gist.github.com/liggitt/bf28159afdaabf52b78ac43d86bc6362, will open upstream go issue